Update spreadsheetbench tools
