+++ environments/jupyter/server/tools/__init__.py
@@ -1,2 +1,1 @@
-from .jupyter import JupyterTool
-from .jupyter_with_record import JupyterToolWithRecord
+from .jupyter import JupyterToolWithRecord
